[Fix] Entities

+++ be/src/main/java/com/niels/homebanking/dto/currency/CreateCurrencyDto.java
@@ -8,13 +8,13 @@ public class CreateCurrencyDto {
 
     private String symbol;
 
-    private int exchange;
+    private long exchange;
 
     public CreateCurrencyDto() {
 
     }
 
-    public CreateCurrencyDto(String name, String iso, String symbol, int exchange) {
+    public CreateCurrencyDto(String name, String iso, String symbol, long exchange) {
         this.name = name;
         this.iso = iso;
         this.symbol = symbol;
@@ -45,11 +45,11 @@ public class CreateCurrencyDto {
         this.symbol = symbol;
     }
 
-    public int getExchange() {
+    public long getExchange() {
         return exchange;
     }
 
-    public void setExchange(int exchange) {
+    public void setExchange(long exchange) {
         this.exchange = exchange;
     }
 }
